created menuitems composable

+++ app/src/main/java/com/example/littlelemon/model/Network.kt
@@ -27,7 +27,7 @@ data class MenuItemNetwork (
         id = id,
         title = title,
         description = description,
-        price = price.toDouble(),
+        price = price,
         image = image,
         category = category
     )
